test: split flaky test CreateMessageTest
