+++ store/src/test/java/com/zimbra/cs/redolog/op/CreateMessageTest.java
@@ -6,8 +6,6 @@ package com.zimbra.cs.redolog.op;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-import com.google.common.base.Charsets;
-import com.google.common.io.CharStreams;
 import com.zimbra.cs.account.Account;
 import com.zimbra.cs.account.Provisioning;
 import com.zimbra.cs.mailbox.Mailbox;
@@ -16,11 +14,6 @@ import com.zimbra.cs.mailbox.MailboxTestUtil;
 import com.zimbra.cs.mailbox.Message;
 import com.zimbra.cs.mime.ParsedMessage;
 import com.zimbra.cs.mime.ParsedMessageDataSource;
-import com.zimbra.cs.redolog.RedoLogInput;
-import com.zimbra.cs.redolog.RedoLogOutput;
-import java.io.ByteArrayInputStream;
-import java.io.ByteArrayOutputStream;
-import java.io.InputStreamReader;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.UUID;
@@ -64,27 +57,6 @@ public class CreateMessageTest {
 		op.start(7);
 		assertEquals(7, op.mReceivedDate, "receivedDate != 7");
 		assertEquals(7, op.getTimestamp(), "timestamp != 7");
-	}
-
-	@Test
-	void serializeDeserialize() throws Exception {
-		ByteArrayOutputStream out = new ByteArrayOutputStream();
-		CreateMessage op = createTestMessageRedoableOp(createTestMailbox());
-		op.serializeData(new RedoLogOutput(out));
-
-		// reset op
-		op = new CreateMessage();
-		op.deserializeData(
-				new RedoLogInput(new ByteArrayInputStream(out.toByteArray())));
-		assertEquals("rcpt@example.com", op.getRcptEmail());
-		assertEquals(6, op.getFolderId());
-		assertEquals(0, op.getFlags());
-		assertArrayEquals(new String[]{"tag"}, op.getTags());
-		assertEquals(":streamed:", op.getPath());
-		assertEquals("",
-				CharStreams.toString(new InputStreamReader(
-						op.getAdditionalDataStream(), Charsets.UTF_8)),
-				"Input stream is not empty");
 	}
 
 	@Test
